Moved recursive parse bigint to its own utils file
